+++ visualizations/4d_heatmaps.py
@@ -10,8 +10,8 @@ import matplotlib.pyplot as plt
 import numpy as np
 
 directory = 'F:/Users/Brendan/Desktop/SolarProject'
-date = '20121018'
-wavelength = 171
+date = '20101208'
+wavelength = 1600
 
 H = np.load('%s/DATA/Output/%s/%i/param.npy' % (directory, date, wavelength))
 #H = np.load('%s/DATA/Output/%s/PCB/%i/param.npy' % (directory, date, wavelength))
@@ -19,11 +19,11 @@ H = np.load('%s/DATA/Output/%s/%i/param.npy' % (directory, date, wavelength))
 titles = [r'Power Law Slope-Coefficient', r'Power Law Index', r'Power Law Tail', r'Gaussian Amplitude', r'Gaussian Location [min]', r'Gaussian Width', 'F-Statistic', r'Gaussian Amplitude Scaled', 'p-Value']
 names = ['slope_coeff', 'index', 'tail', 'gauss_amp', 'gauss_loc', 'gauss_wid', 'f_test', 'gauss_amp_scaled', 'p_value']
 
-p1 = 1
-p2 = 4
+p1 = 5
+p2 = 2
 
-param1 = H[p1,0:,0:] # for 20160520
-param2 = H[p2,0:,0:] # for 20160520
+param1 = H[p1,50:-50,70:-70] # for 20160520
+param2 = H[p2,50:-50,70:-70] # for 20160520
 
 #param1 = H[p1,180:-175,350:-80] # for 20140818
 #param2 = H[p2,180:-175,350:-80] # for 20140818